requirements.txt
django==4.0
psycopg2-binary==2.9.2
requests>=2.26.0
xmlsec==1.3.12
pysaml2>=4.5.0
is-safe-url>=1.0
django-rest-auth==0.9.5
django-filter>=21.1
djangorestframework==3.13.1
djangorestframework-jwt>=1.11.0
kubernetes==21.7.0
boto3>=1.20.39